[Model] Consolidate Deepseek-MoE implementation with DeepSeek-v2 (#28101)
